added string equal

+++ data_structures/test_array_list.c
@@ -16,7 +16,7 @@ main(int argc, char* argv[])
         slice_t buf_slice = make_slice((void *)input_array, sizeof(input_array)); 
         int ret = array_list_append_slice_fn(&gpa, &float_list, buf_slice);
         float *ptr = (float *)(float_list.ptr);
-        for (int i = 0; i < float_list.len; i++){
+        for (size_t i = 0; i < float_list.len; i++){
             printf("float_list[%d]: %f\n", i, ptr[i]);
         }
     }
